+++ src/main/kotlin/icu/windea/pls/cwt/codeStyle/CwtLanguageCodeStyleSettingsProvider.kt
@@ -22,7 +22,7 @@ class CwtLanguageCodeStyleSettingsProvider : LanguageCodeStyleSettingsProvider()
 		indentOptions.INDENT_SIZE = 4
 		indentOptions.CONTINUATION_INDENT_SIZE = 4
 		indentOptions.KEEP_INDENTS_ON_EMPTY_LINES = true
-		indentOptions.USE_TAB_CHARACTER = true
+		indentOptions.USE_TAB_CHARACTER = false
 		commonSettings.LINE_COMMENT_AT_FIRST_COLUMN = false
 		commonSettings.LINE_COMMENT_ADD_SPACE = false
 	}