+++ src/main/kotlin/icu/windea/pls/cwt/formatter/CwtBlock.kt
@@ -18,6 +18,7 @@ class CwtBlock(
 	private val settings: CodeStyleSettings
 ) : AbstractBlock(node, createWrap(), createAlignment()) {
 	companion object {
+		private val MEMBERS = TokenSet.create(PROPERTY, VALUE, BOOLEAN, INT, FLOAT, STRING, BLOCK)
 		private val SEPARATORS = TokenSet.create(EQUAL_SIGN, NOT_EQUAL_SIGN)
 		private val SHOULD_INDENT_PARENT_TYPES = TokenSet.create(BLOCK)
 		private val SHOULD_INDENT_TYPES = TokenSet.create(PROPERTY, VALUE, BOOLEAN, INT, FLOAT, STRING, BLOCK, COMMENT, DOCUMENTATION_COMMENT, OPTION_COMMENT)
@@ -35,10 +36,11 @@ class CwtBlock(
 			//变量声明分隔符周围的空格，属性分隔符周围的空格
 			val customSettings = settings.getCustomSettings(CwtCodeStyleSettings::class.java)
 			return SpacingBuilder(settings, CwtLanguage)
+				.between(MEMBERS, MEMBERS).spaces(1, true) //属性/值之间需要有空格或者换行
 				.aroundInside(SEPARATORS, OPTION).spaceIf(customSettings.SPACE_AROUND_OPTION_SEPARATOR) //等号、不等号周围按情况可能需要空格
 				.aroundInside(SEPARATORS, PROPERTY).spaceIf(customSettings.SPACE_AROUND_PROPERTY_SEPARATOR) //等号、不等号周围按情况可能需要空格
 				.between(LEFT_BRACE, RIGHT_BRACE).none() //花括号之间总是不需要空格
-				.withinPair(LEFT_BRACE,RIGHT_BRACE).spaceIf( customSettings.SPACE_WITHIN_BRACES) //花括号内侧如果非换行按情况可能需要空格
+				.withinPair(LEFT_BRACE,RIGHT_BRACE).spaceIf( customSettings.SPACE_WITHIN_BRACES, true) //花括号内侧如果非换行按情况可能需要空格
 		}
 	}