+++ src/main/kotlin/icu/windea/pls/script/formatter/ParadoxScriptBlock.kt
@@ -9,6 +9,7 @@ import com.intellij.psi.formatter.common.*
 import com.intellij.psi.tree.*
 import com.intellij.util.*
 import icu.windea.pls.core.*
+import icu.windea.pls.cwt.formatter.*
 import icu.windea.pls.script.*
 import icu.windea.pls.script.codeStyle.*
 import icu.windea.pls.script.psi.ParadoxScriptElementTypes.*
@@ -18,6 +19,7 @@ class ParadoxScriptBlock(
 	private val settings: CodeStyleSettings,
 ) : AbstractBlock(node, createWrap(), createAlignment()) {
 	companion object {
+		private val MEMBERS = TokenSet.create(SCRIPTED_VARIABLE, PROPERTY, BOOLEAN, INT, FLOAT, STRING, COLOR, INLINE_MATH, PARAMETER, BLOCK, PARAMETER_CONDITION)
 		private val SEPARATORS = TokenSet.create(EQUAL_SIGN, NOT_EQUAL_SIGN, LT_SIGN, GT_SIGN, LE_SIGN, GE_SIGN)
 		private val INLINE_MATH_OPERATORS = TokenSet.create(PLUS_SIGN, MINUS_SIGN, TIMES_SIGN, DIV_SIGN, MOD_SIGN, LABS_SIGN, RABS_SIGN, LP_SIGN, RP_SIGN)
 		private val SHOULD_INDENT_PARENT_TYPES = TokenSet.create(BLOCK, PARAMETER_CONDITION)
@@ -36,17 +38,18 @@ class ParadoxScriptBlock(
 			//变量声明分隔符周围的空格，属性分隔符周围的空格
 			val customSettings = settings.getCustomSettings(ParadoxScriptCodeStyleSettings::class.java)
 			return SpacingBuilder(settings, ParadoxScriptLanguage)
+				.between(MEMBERS, MEMBERS).spaces(1, true) //封装变量/属性/值/参数条件块之间需要有空格或者换行
 				.aroundInside(SEPARATORS, SCRIPTED_VARIABLE).spaceIf(customSettings.SPACE_AROUND_SCRIPTED_VARIABLE_SEPARATOR) //间隔符周围按情况可能需要空格
 				.aroundInside(SEPARATORS, PROPERTY).spaceIf(customSettings.SPACE_AROUND_PROPERTY_SEPARATOR) //间隔符周围按情况可能需要空格
 				.around(INLINE_MATH_OPERATORS).spaceIf(customSettings.SPACE_AROUND_INLINE_MATH_OPERATOR) //内联数学表达式操作符周围按情况可能需要空格
 				.between(LEFT_BRACE, RIGHT_BRACE).none()//花括号之间总是不需要空格
-				.withinPair(LEFT_BRACE, RIGHT_BRACE).spaceIf(customSettings.SPACE_WITHIN_BRACES) //花括号内侧如果非换行按情况可能需要空格
+				.withinPair(LEFT_BRACE, RIGHT_BRACE).spaceIf(customSettings.SPACE_WITHIN_BRACES, true) //花括号内侧如果非换行按情况可能需要空格
 				.between(NESTED_LEFT_BRACKET, NESTED_RIGHT_BRACKET).none() //参数条件表达式如果为空则不需要空格（尽管这是语法错误）
-				.withinPair(NESTED_LEFT_BRACKET, NESTED_RIGHT_BRACKET).spaceIf(customSettings.SPACE_WITHIN_PARAMETER_CONDITION_EXPRESSION_BRACKETS) //参数条件表达式内侧如果非换行按情况可能需要空格
+				.withinPair(NESTED_LEFT_BRACKET, NESTED_RIGHT_BRACKET).spaceIf(customSettings.SPACE_WITHIN_PARAMETER_CONDITION_EXPRESSION_BRACKETS, true) //参数条件表达式内侧如果非换行按情况可能需要空格
 				.between(NESTED_RIGHT_BRACKET, RIGHT_BRACKET).none() //参数条件代码块如果为空则不需要空格
-				.withinPair(NESTED_RIGHT_BRACKET, RIGHT_BRACKET).spaceIf(customSettings.SPACE_WITHIN_PARAMETER_CONDITION_BRACKETS) //参数条件代码块内侧如果非换行按情况可能需要空格
+				.withinPair(NESTED_RIGHT_BRACKET, RIGHT_BRACKET).spaceIf(customSettings.SPACE_WITHIN_PARAMETER_CONDITION_BRACKETS, true) //参数条件代码块内侧如果非换行按情况可能需要空格
 				.between(INLINE_MATH_START, INLINE_MATH_END).none() //内联数字表达式如果为空则不需要空格（尽管这是语法错误）
-				.withinPair(INLINE_MATH_START, INLINE_MATH_END).spaceIf(customSettings.SPACE_WITHIN_INLINE_MATH_BRACKETS) //内联数学表达式内侧如果非换行按情况可能需要空格
+				.withinPair(INLINE_MATH_START, INLINE_MATH_END).spaceIf(customSettings.SPACE_WITHIN_INLINE_MATH_BRACKETS, true) //内联数学表达式内侧如果非换行按情况可能需要空格
 		}
 	}